GSOI-BC/app/requirements.txt
kafka-python
datime
json
fastapi
starlette.responses
pydantic
uvicorn
uuid
threading
time
random
pycrypto
base58
requests